+++ tree_queries/models.py
@@ -1,12 +1,15 @@
+from __future__ import unicode_literals
+
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from tree_queries.query import TreeQuerySet
+from .fields import TreeNodeForeignKey
+from .query import TreeQuerySet
 
 
 class TreeNode(models.Model):
-    parent = models.ForeignKey(
+    parent = TreeNodeForeignKey(
         "self",
         blank=True,
         null=True,